move main/provisonal tests to e2e
